chore: Separate command
