+++ main/main.go
@@ -1,8 +1,6 @@
 package main
 
-import (
-	"github.com/LarsNieuwenhuizen/loadbalancer/cmd"
-)
+import "github.com/LarsNieuwenhuizen/loadbalancer/cmd"
 
 func main() {
 	cmd.Execute()
